Mode dirty cache to its own file
